Rename media model module and add validated recording-state publisher
Rename device_viewer/models/media_capture_model.py -> media.py and add
RecordingActiveState / RecordingStatePublisher (a ValidatedTopicPublisher)
so recording-state messages carry a typed StrictBool payload instead of a
raw true/false string. message_listener now decodes via
RecordingActiveState.model_validate_json.

Relocate DEVICE_VIEWER_RECORDING_STATE to device_viewer.consts (its
canonical home) with the publisher instance; UITopics and protocol_grid
re-export from there. Update all import sites for the module rename.
