+++ device_viewer/views/camera_control_view/widget.py
@@ -25,7 +25,7 @@ from microdrop_application.helpers import get_current_experiment_directory
 from microdrop_utils.dramatiq_pub_sub_helpers import publish_message
 from microdrop_utils.pyside_helpers import MarqueeComboBox
 from microdrop_utils.v4l2_fps_getter import get_video_inputs, LinuxCameraDeviceContainer
-from protocol_grid.consts import DEVICE_VIEWER_RECORDING_STATE
+from ...consts import DEVICE_VIEWER_RECORDING_STATE, device_viewer_recording_state_publisher
 
 from device_viewer.views.camera_control_view.preferences import CameraPreferences
 from microdrop_style.helpers import get_complete_stylesheet, is_dark_mode
@@ -39,7 +39,7 @@ from ...utils.camera import (
     get_transformed_frame,
     ImageSaver,
 )
-from ...models.media_capture_model import MediaType
+from ...models.media import MediaType
 
 from logger.logger_service import get_logger
 logger = get_logger(__name__)
@@ -739,7 +739,7 @@ class CameraControlWidget(QWidget):
         self.recorder.start(
             _recording_file_path, _resolution, fps
         )
-        publish_message(topic=DEVICE_VIEWER_RECORDING_STATE, message="true")
+        device_viewer_recording_state_publisher.publish(state=True)
 
     def video_record_stop(self):
         logger.info("Stopping video recorder...")
@@ -748,7 +748,7 @@ class CameraControlWidget(QWidget):
     @Slot(str)
     def handle_recording_error(self, error_msg):
         logger.error(f"Recording Error: {error_msg}")
-        publish_message(topic=DEVICE_VIEWER_RECORDING_STATE, message="false")
+        device_viewer_recording_state_publisher.publish(state=False)
         error(
             self,
             "<b>Error</b>: Cannot continue to record video<br><br>Exception raised while recording video.",
@@ -758,7 +758,7 @@ class CameraControlWidget(QWidget):
 
     @Slot(str)
     def handle_recording_stopped(self, recording_output_path):
-        publish_message(topic=DEVICE_VIEWER_RECORDING_STATE, message="false")
+        device_viewer_recording_state_publisher.publish(state=False)
         if not self._camera_state_pre_recording:
             # turn off camera if we need to
             if self.camera.isActive():
