+++ examples/tests/test_messaging_api.py
@@ -153,9 +153,9 @@ class TestUITopicsMatchConsts:
             DEVICE_VIEWER_STATE_CHANGED, PROTOCOL_GRID_DISPLAY_STATE,
             CALIBRATION_DATA, DEVICE_VIEWER_SCREEN_CAPTURE,
             DEVICE_VIEWER_SCREEN_RECORDING, DEVICE_VIEWER_CAMERA_ACTIVE,
-            DEVICE_VIEWER_MEDIA_CAPTURED, DEVICE_VIEWER_RECORDING_STATE,
-            ROUTES_EXECUTING,
+            DEVICE_VIEWER_MEDIA_CAPTURED, ROUTES_EXECUTING,
         )
+        from device_viewer.consts import DEVICE_VIEWER_RECORDING_STATE
         assert UITopics.DEVICE_VIEWER_STATE_CHANGED == DEVICE_VIEWER_STATE_CHANGED
         assert UITopics.PROTOCOL_GRID_DISPLAY_STATE == PROTOCOL_GRID_DISPLAY_STATE
         assert UITopics.CALIBRATION_DATA == CALIBRATION_DATA
