+++ protocol_grid/consts.py
@@ -25,11 +25,8 @@ PROTOCOL_GRID_LISTENER_NAME = f"{PKG}_listener"
 # Re-exported from device_viewer.consts for back-compat (canonical home moved there).
 # Safe to remove this re-export block once PPT-9 deletes the protocol_grid plugin.
 from device_viewer.consts import (
-    DEVICE_VIEWER_SCREEN_CAPTURE,
-    DEVICE_VIEWER_SCREEN_RECORDING,
-    DEVICE_VIEWER_CAMERA_ACTIVE,
     DEVICE_VIEWER_MEDIA_CAPTURED,
-    CALIBRATION_DATA,
+    CALIBRATION_DATA, DEVICE_VIEWER_RECORDING_STATE,
 )
 
 DEVICE_VIEWER_STATE_CHANGED = "ui/device_viewer/state_changed"
@@ -37,7 +34,6 @@ STEP_PARAMS_COMMIT = "ui/device_viewer/step_params_commit"
 PROTOCOL_GRID_DISPLAY_STATE = "ui/protocol_grid/display_state"
 PROTOCOL_RUNNING = "microdrop/protocol_running"
 ROUTES_EXECUTING = "ui/device_viewer/routes_executing"
-DEVICE_VIEWER_RECORDING_STATE = "ui/device_viewer/recording_state"
 
 ACTOR_TOPIC_DICT = {
     PROTOCOL_GRID_LISTENER_NAME: [
